feat: 支持批量操作多个频道
