+++ spectre-core/src/main/kotlin/io/github/vudsen/spectre/core/vo/BatchChannelsRequestVO.kt
@@ -1,0 +1,8 @@
+package io.github.vudsen.spectre.core.vo
+
+import jakarta.validation.constraints.NotEmpty
+
+class BatchChannelsRequestVO {
+    @NotEmpty
+    var channelIds: List<String> = emptyList()
+}
