+++ spectre-core/src/main/kotlin/io/github/vudsen/spectre/core/vo/BatchExecRequestVO.kt
@@ -1,0 +1,11 @@
+package io.github.vudsen.spectre.core.vo
+
+import jakarta.validation.constraints.NotEmpty
+
+class BatchExecRequestVO {
+    @NotEmpty
+    var channelIds: List<String> = emptyList()
+
+    @NotEmpty
+    var command: String = ""
+}
