+++ spectre-frontend/src/pages/channel/BatchChannelCreate.tsx
@@ -1,0 +1,9 @@
+interface BatchChannelCreateProps {
+  channels: { treeNodeId: string; bundleId: string; runtimeNodeId: string }[]
+}
+
+const BatchChannelCreate: React.FC<BatchChannelCreateProps> = (props) => {
+  // TODO
+}
+
+export default BatchChannelCreate
